Add S3EndpointMiddleware instead
